4.1. Playback State
The playbackState attribute specifies the declared playback state from the browsing context. The state is combined with the guessed playback state to compute the actual playback state, which is a finalized state and will be used for play and pause actions.

The actual playback state is computed in the following way:

If the declared playback state is playing, return playing.
Otherwise, return the guessed playback state.
The playbackState attribute could be useful when the page wants to do some preparation steps when the media is paused but it allows the preparation steps to be interrupted by pause action. See Setting playbackState for example.

When the actual playback state of the active media session changes, the user agent MUST run the media session actions update algorithm.